466: create draw/path.js
